# docs/openapi/phase2-contract.yaml
openapi: 3.1.0
info:
  title: RankWoven Phase 2 Contract
  version: 0.1.0
  description: >-
    Phase 2A foundation endpoints. All write operations require a Bearer JWT and
    an Idempotency-Key. Provider and CMS side effects are executed by workers.
servers:
  - url: https://api.rankwoven.com
paths:
  /api/v1/tasks/{taskId}:
    get:
      summary: Read a workspace-scoped task
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { $ref: '#/components/responses/TaskResponse' }
        '401': { $ref: '#/components/responses/AuthError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/tasks/{taskId}/cancel:
    post:
      summary: Cancel a queued or running task
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { $ref: '#/components/responses/TaskResponse' }
        '202': { $ref: '#/components/responses/TaskResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /api/v1/tasks/dead-letter:
    get:
      summary: List workspace-scoped dead-letter tasks
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/DeadLetterTasksResponse' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/tasks/{taskId}/replays:
    post:
      summary: Create an auditable replacement for a dead-letter task
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeadLetterActionRequest' }
      responses:
        '202': { $ref: '#/components/responses/DeadLetterReplayResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/tasks/{taskId}/dead-letter/ignore:
    post:
      summary: Record an ignore disposition for a dead-letter task
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeadLetterActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/DeadLetterIgnoreResponse' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/usage:
    get:
      summary: Read workspace usage ledger totals
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageResponse' }
  /api/v1/admin/ai-gateway/models:
    get:
      summary: List the server-side Breakout model catalog
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Model metadata only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelsResponse' }
  /api/v1/admin/ai-gateway/models/sync:
    post:
      summary: Queue a Breakout model catalog sync
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/TaskAccepted' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/admin/ai-gateway/profiles:
    get:
      summary: Read current gateway model profile assignments
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile assignments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfilesResponse' }
  /api/v1/admin/ai-gateway/profiles/{profileKey}:
    put:
      summary: Assign a verified Breakout model to a profile
      security: [{ bearerAuth: [] }]
      parameters:
        - name: profileKey
          in: path
          required: true
          schema:
            type: string
            enum: [text.default, text.high_quality, text.batch, embedding.default, image.default]
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelId]
              properties:
                modelId: { type: string }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    TaskResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [task]
          properties: { task: { $ref: '#/components/schemas/Task' } }
...
    UsageResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [reservedUnits, activeReservedUnits, finalizedUnits, releasedUnits, reservedCost, finalizedCost, releasedCost]
          properties:
            reservedUnits: { type: number }
            activeReservedUnits: { type: number }
            finalizedUnits: { type: number }
            releasedUnits: { type: number }
            reservedCost: { type: number }
            finalizedCost: { type: number }
            releasedCost: { type: number }
    DeadLetterActionRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
...
    DeadLetterTasksResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [items, pagination]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Task' } }
            pagination: { $ref: '#/components/schemas/Pagination' }
    DeadLetterReplayResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [task, action]
          properties:
            task: { $ref: '#/components/schemas/Task' }
            action: { $ref: '#/components/schemas/DeadLetterAction' }
    DeadLetterIgnoreResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [action]
          properties:
            action: { $ref: '#/components/schemas/DeadLetterAction' }
    ModelsResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          properties:
            models:
              type: array
              items:
                type: object
                required: [gateway, modelId, supportedEndpointTypes, capabilityStatus, catalogHash]
                properties:
                  gateway: { type: string, enum: [wenwen] }
                  modelId: { type: string }
                  ownedBy: { type: string, nullable: true }
                  supportedEndpointTypes: { type: array, items: { type: string, enum: [chat, embedding, image] } }
                  capabilityStatus: { type: string, enum: [pending, verified, unavailable] }
                  catalogHash: { type: string }
                  verifiedAt: { type: string, format: date-time, nullable: true }
    ProfilesResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [profiles]
          properties:
            profiles:
              type: array
              items:
                type: object
                required: [profileKey, modelId, status]
                properties:
                  profileKey: { type: string }
                  modelId: { type: string, nullable: true }
                  status: { type: string, enum: [unconfigured, pending, active] }
    ProfileResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          required: [profile]
          properties:
            profile:
              type: object
              required: [gateway, profileKey, modelId, version, status, updatedAt]
              properties:
                gateway: { type: string, enum: [wenwen] }
                profileKey: { type: string }
                modelId: { type: string }
                version: { type: integer, minimum: 1 }
                status: { type: string, enum: [active, disabled] }
                updatedAt: { type: string, format: date-time }
                updatedBy: { type: string, nullable: true }
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    Error:
      type: object
      required: [success, message, error]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        error:
          type: object
          required: [code]
          properties:
            code: { type: string }
            details: {}
  responses:
    TaskResponse:
      description: Task response
      content: { application/json: { schema: { $ref: '#/components/schemas/TaskResponse' } } }
    TaskAccepted:
      description: Task accepted
      content: { application/json: { schema: { $ref: '#/components/schemas/TaskAcceptedResponse' } } }
    DeadLetterTasksResponse:
      description: Paginated dead-letter tasks in the current workspace
      content: { application/json: { schema: { $ref: '#/components/schemas/DeadLetterTasksResponse' } } }
    DeadLetterReplayResponse:
      description: Replacement task and immutable replay disposition
      content: { application/json: { schema: { $ref: '#/components/schemas/DeadLetterReplayResponse' } } }
    DeadLetterIgnoreResponse:
      description: Immutable ignore disposition
      content: { application/json: { schema: { $ref: '#/components/schemas/DeadLetterIgnoreResponse' } } }
    AuthError:
      description: Authentication failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ForbiddenError:
      description: Insufficient role
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFoundError:
      description: Workspace-scoped resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ConflictError:
      description: Idempotency or task state conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ProviderUnavailable:
      description: Provider or capability unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...